requirements.txt
tensorflow==1.14.0
numpy==1.16.5
scipy==1.3.1
matplotlib==3.0.3
pillow==5.3.0
scikit-image==0.15.0
google-api-python-client==1.7.11
apiclient==1.0.4
google-cloud==0.34.0
google-cloud-storage==1.19.1